django_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
requests==2.31.0
python-decouple==3.8
django-cors-headers==4.3.1